move plotting code
